+++ src/client/cypress/e2e/editor/casing.cy.js
@@ -71,6 +71,11 @@ describe("Casing crud tests", () => {
     evaluateDisplayValue("casingElements.0.innerDiameter", "3");
     evaluateDisplayValue("casingElements.0.outerDiameter", "4");
 
+    // add casing element and verify fromDepth is set to previous toDepth
+    addItem("addCasingElement");
+    evaluateDisplayValue("casingElements.1.fromDepth", "10");
+    cy.get('[data-cy="casingElements.1.delete"]').click();
+
     // update casing
     startEditing();
     cy.wait("@codelist_GET");